+++ AWSPinpointTests/AWSPinpointTargetingClientTests.m
@@ -22,7 +22,11 @@
 #import "AWSPinpointNotificationManager.h"
 
 NSString *const AWSPinpointTargetingClientErrorDomain = @"com.amazonaws.AWSPinpointAnalyticsClientErrorDomain";
-
+NSString *const AWSPinpointTargetingClientTestsName = @"AWSPinpointTargetingClientTests";
+NSString *const AWSPinpointEndpointAttributesKey = @"AWSPinpointEndpointAttributesKey";
+NSString *const AWSPinpointEndpointMetricsKey = @"AWSPinpointEndpointMetricsKey";
+NSString *const AWSPinpointEndpointProfileKey = @"AWSPinpointEndpointProfileKey";
+NSString *const AWSDeviceToken = @"com.amazonaws.AWSDeviceTokenKey";
 static NSString *userId;
 
 @interface AWSPinpointTargetingClientTests : XCTestCase
@@ -52,13 +56,11 @@ static NSString *userId;
     [super setUp];
 
     [AWSTestUtility setupCognitoCredentialsProvider];
-    [[NSUserDefaults standardUserDefaults] removeSuiteNamed:@"AWSPinpointTargetingClientTests"];
-    self.userDefaults = [[NSUserDefaults alloc] initWithSuiteName:@"AWSPinpointTargetingClientTests"];
-
-    [self initializeMockApplicationWithOptOut:YES];
-    [self initializePinpointWithConfiguration:[self getDefaultAWSPinpointConfiguration] forceCreate:NO];
-    [self.userDefaults removeObjectForKey:@"AWSPinpointEndpointAttributesKey"];
-    [self.userDefaults removeObjectForKey:@"AWSPinpointEndpointMetricsKey"];
+    [[NSUserDefaults standardUserDefaults] removeSuiteNamed:AWSPinpointTargetingClientTestsName];
+    self.userDefaults = [[NSUserDefaults alloc] initWithSuiteName:AWSPinpointTargetingClientTestsName];
+    [self.userDefaults removeObjectForKey:AWSPinpointEndpointAttributesKey];
+    [self.userDefaults removeObjectForKey:AWSPinpointEndpointMetricsKey];
+    [self.userDefaults removeObjectForKey:AWSDeviceTokenKey];
     [self.userDefaults synchronize];
 }
 
@@ -79,19 +81,22 @@ static NSString *userId;
 - (AWSPinpointConfiguration *)getAWSPinpointConfigurationWithOptOut:(BOOL)optOut {
     AWSPinpointConfiguration *configuration = [self getDefaultAWSPinpointConfiguration];
 
-    [self setApplicationLevelOptOut:configuration withOptOut:optOut];
+    configuration.isApplicationLevelOptOut = ^BOOL{
+        return optOut;
+    };
 
     return configuration;
 }
 
-- (void)setApplicationLevelOptOut:(AWSPinpointConfiguration *)configuration withOptOut:(BOOL)optOut {
-    configuration.isApplicationLevelOptOut = ^BOOL{
-        return optOut;
+- (void)setDeviceTokenInUserDefaults {
+    const unsigned char currentTokenBytes[] = {
+        0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
+        0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f
     };
-}
-
-- (void)setApplicationLevelOptOut:(BOOL)optOut {
-    [self initializePinpointWithConfiguration:[self getAWSPinpointConfigurationWithOptOut:optOut]];
+    NSData *currentTokenData = [[NSData alloc] initWithBytes:currentTokenBytes length:16];
+    [self.userDefaults removeObjectForKey:AWSDeviceTokenKey];
+    [self.userDefaults setObject:currentTokenData forKey:AWSDeviceTokenKey];
+    [self.userDefaults synchronize];
 }
 
 - (void)initializePinpointWithConfiguration:(AWSPinpointConfiguration *)configuration forceCreate:(BOOL)forceCreate {
@@ -107,27 +112,15 @@ static NSString *userId;
     self.pinpoint = [AWSPinpoint pinpointWithConfiguration:configuration];
 }
 
-- (void)initializePinpointWithConfiguration:(AWSPinpointConfiguration *)configuration {
-    [self initializePinpointWithConfiguration:configuration forceCreate:YES];
-}
-
-- (void)initializeMockApplicationWithOptOut:(BOOL)optOut {
+- (void)initializeMockApplicationWithRemoteNotifications:(BOOL)withRemoteNotifications {
     id mockApplication = OCMClassMock([UIApplication class]);
 
     OCMStub([mockApplication sharedApplication]).andReturn(mockApplication);
 
-    [self setSystemLevelOptOut:mockApplication withOptOut:optOut];
-}
-
-- (void)setSystemLevelOptOut:(UIApplication *)application withOptOut:(BOOL)optOut {
-    UIUserNotificationType notificationType = optOut ? UIUserNotificationTypeNone : UIUserNotificationTypeAlert;
+    UIUserNotificationType notificationType = withRemoteNotifications ? UIUserNotificationTypeAlert : UIUserNotificationTypeNone;
     UIUserNotificationSettings *notificationSettings = [UIUserNotificationSettings settingsForTypes:notificationType categories:nil];
-    OCMStub([application currentUserNotificationSettings]).andReturn(notificationSettings);
-    OCMStub([application isRegisteredForRemoteNotifications]).andReturn(optOut == NO);
-}
-
-- (void)setSystemLevelOptOut:(BOOL)optOut {
-    [self initializeMockApplicationWithOptOut:optOut];
+    OCMStub([mockApplication currentUserNotificationSettings]).andReturn(notificationSettings);
+    OCMStub([mockApplication isRegisteredForRemoteNotifications]).andReturn(withRemoteNotifications);
 }
 
 - (void)testConstructors {
@@ -141,8 +134,9 @@ static NSString *userId;
 }
 
 - (void)testEndpointProfileInformationPersistence {
+    [self initializePinpointWithConfiguration:[self getDefaultAWSPinpointConfiguration] forceCreate:NO];
     NSString *dummyAppId = @"dummyAppId";
-    [self.pinpoint.configuration.userDefaults removeObjectForKey:@"AWSPinpointEndpointProfileKey"];
+    [self.pinpoint.configuration.userDefaults removeObjectForKey:AWSPinpointEndpointProfileKey];
     [self.pinpoint.configuration.userDefaults synchronize];
     AWSPinpointEndpointProfile *endpointProfile = [self.pinpoint.targetingClient currentEndpointProfile];
     endpointProfile.user.userId = userId;
@@ -150,7 +144,7 @@ static NSString *userId;
         XCTAssertNil(task.error);
         return nil;
     }] waitUntilFinished];
-    XCTAssertNotNil([self.pinpoint.configuration.userDefaults objectForKey:@"AWSPinpointEndpointProfileKey"]);
+    XCTAssertNotNil([self.pinpoint.configuration.userDefaults objectForKey:AWSPinpointEndpointProfileKey]);
     AWSPinpointEndpointProfile *profile = [self.pinpoint.targetingClient currentEndpointProfile];
     XCTAssertTrue([profile.user.userId isEqualToString:userId]);
     AWSPinpoint *pinpoint = [AWSPinpoint pinpointWithConfiguration:self.configuration];
@@ -249,49 +243,57 @@ static NSString *userId;
     [userDefaults synchronize];
 }
 
-- (void)testCurrentProfileWithSystemOptOutAndApplicationOptOut {
-    [self setSystemLevelOptOut:YES];
-    [self setApplicationLevelOptOut:YES];
+- (void)testCurrentProfileReturnsOptOutAll {
+    [self initializePinpointWithConfiguration:[self getAWSPinpointConfigurationWithOptOut:NO] forceCreate:YES];
+    AWSPinpointEndpointProfile *profile = [self.pinpoint.targetingClient currentEndpointProfile];
+    XCTAssertTrue([profile.optOut isEqualToString:@"ALL"]);
+}
+
+- (void)testCurrentProfileReturnsOptOutAllWhenNotificationsEnabledAndDeviceTokenNotSet {
+    [self initializePinpointWithConfiguration:[self getAWSPinpointConfigurationWithOptOut:NO] forceCreate:YES];
+    [self initializeMockApplicationWithRemoteNotifications:YES];
 
     AWSPinpointEndpointProfile *profile = [self.pinpoint.targetingClient currentEndpointProfile];
     XCTAssertTrue([profile.optOut isEqualToString:@"ALL"]);
 }
 
-- (void)testCurrentProfileWithSystemOptOutAndApplicationOptIn {
-    [self setSystemLevelOptOut:YES];
-    [self setApplicationLevelOptOut:NO];
+- (void)testCurrentProfileReturnsOptOutAllWhenNotificationsDisabledAndDeviceTokenSet {
+    [self initializePinpointWithConfiguration:[self getAWSPinpointConfigurationWithOptOut:NO] forceCreate:YES];
+    [self initializeMockApplicationWithRemoteNotifications:NO];
+    [self setDeviceTokenInUserDefaults];
 
     AWSPinpointEndpointProfile *profile = [self.pinpoint.targetingClient currentEndpointProfile];
-
     XCTAssertTrue([profile.optOut isEqualToString:@"ALL"]);
 }
 
-- (void)testCurrentProfileWithSystemOptOutAndApplicationOptOutUnset {
-    [self setSystemLevelOptOut:YES];
+- (void)testCurrentProfileReturnsOptOutNoneWhenNotificationsEnabledAndDeviceTokenSet {
+    [self initializePinpointWithConfiguration:[self getAWSPinpointConfigurationWithOptOut:NO] forceCreate:YES];
+    [self initializeMockApplicationWithRemoteNotifications:YES];
+    [self setDeviceTokenInUserDefaults];
 
     AWSPinpointEndpointProfile *profile = [self.pinpoint.targetingClient currentEndpointProfile];
+    XCTAssertTrue([profile.optOut isEqualToString:@"NONE"]);
+}
 
+- (void)testCurrentProfileReturnsOptOutAllForApplicationLevelOptOut {
+    [self initializePinpointWithConfiguration:[self getAWSPinpointConfigurationWithOptOut:YES] forceCreate:YES];
+    [self initializeMockApplicationWithRemoteNotifications:YES];
+    [self setDeviceTokenInUserDefaults];
+
+    AWSPinpointEndpointProfile *profile = [self.pinpoint.targetingClient currentEndpointProfile];
     XCTAssertTrue([profile.optOut isEqualToString:@"ALL"]);
 }
 
-- (void)testCurrentProfileWithSystemOptInAndApplicationOptOut {
-    [self setSystemLevelOptOut:NO];
-    [self setApplicationLevelOptOut:YES];
-
-    AWSPinpointEndpointProfile *profile = [self.pinpoint.targetingClient currentEndpointProfile];
-
-    XCTAssertTrue([profile.optOut isEqualToString:@"ALL"]);
-}
-
-- (void) testCurrentProfileWithSystemOptInAndApplicationOptOutBackgroundThread {
+- (void) testCurrentProfileReturnsOptOutNoneWithNotificationsEnabledAndDeviceTokenSetOnBackgroundThread {
     __block XCTestExpectation *expectation = [self expectationWithDescription:@"Test finished running."];
     
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0) , ^{
-        [self setSystemLevelOptOut:NO];
-        [self setApplicationLevelOptOut:YES];
+        [self initializePinpointWithConfiguration:[self getAWSPinpointConfigurationWithOptOut:NO] forceCreate:YES];
+        [self initializeMockApplicationWithRemoteNotifications:YES];
+        [self setDeviceTokenInUserDefaults];
 
         AWSPinpointEndpointProfile *profile = [self.pinpoint.targetingClient currentEndpointProfile];
-        XCTAssertTrue([profile.optOut isEqualToString:@"ALL"]);
+        XCTAssertTrue([profile.optOut isEqualToString:@"NONE"]);
         [expectation fulfill];
         expectation = nil;
     });
@@ -301,22 +303,24 @@ static NSString *userId;
     }];
 }
 
-- (void)testCurrentProfileWithSystemOptInAndApplicationOptIn {
-    [self setSystemLevelOptOut:NO];
-    [self setApplicationLevelOptOut:NO];
+- (void) testCurrentProfileReturnsOptOutAllWithNotificationsEnabledAndDeviceTokenNotSetOnBackgroundThread {
+    __block XCTestExpectation *expectation = [self expectationWithDescription:@"Test finished running."];
 
-    AWSPinpointEndpointProfile *profile = [self.pinpoint.targetingClient currentEndpointProfile];
+    dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0) , ^{
+        [self initializePinpointWithConfiguration:[self getAWSPinpointConfigurationWithOptOut:NO] forceCreate:YES];
+        [self initializeMockApplicationWithRemoteNotifications:YES];
 
-    XCTAssertTrue([profile.optOut isEqualToString:@"NONE"]);
+        AWSPinpointEndpointProfile *profile = [self.pinpoint.targetingClient currentEndpointProfile];
+        XCTAssertTrue([profile.optOut isEqualToString:@"ALL"]);
+        [expectation fulfill];
+        expectation = nil;
+    });
+
+    [self waitForExpectationsWithTimeout:5 handler:^(NSError * _Nullable error) {
+        XCTAssertNil(error);
+    }];
 }
 
-- (void)testCurrentProfileWithSystemOptInAndApplicationOptOutUnset {
-    [self setSystemLevelOptOut:NO];
-
-    AWSPinpointEndpointProfile *profile = [self.pinpoint.targetingClient currentEndpointProfile];
-
-    XCTAssertTrue([profile.optOut isEqualToString:@"NONE"]);
-}
 - (void)testCurrentProfile {
     [self validateCurrentProfile:NO forAppId:@"testCurrentProfileForAPNS"];
     [self validateCurrentProfile:YES forAppId:@"testCurrentProfileForAPNSSandbox"];
@@ -378,6 +382,7 @@ static NSString *userId;
 }
 
 - (void)testUpdateEndpointProfileWithProfile {
+    [self initializePinpointWithConfiguration:[self getDefaultAWSPinpointConfiguration] forceCreate:NO];
     AWSPinpointEndpointProfile *endpointProfile = [self.pinpoint.targetingClient currentEndpointProfile];
     endpointProfile.user.userId = userId;
     [[[self.pinpoint.targetingClient updateEndpointProfile:endpointProfile] continueWithBlock:^id _Nullable(AWSTask * _Nonnull task) {
@@ -447,6 +452,7 @@ static NSString *userId;
 }
 
 - (void) testGlobalAttributeValidation {
+    [self initializePinpointWithConfiguration:[self getDefaultAWSPinpointConfiguration] forceCreate:NO];
     @try {
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wnonnull"
@@ -547,6 +553,7 @@ static NSString *userId;
 }
 
 - (void) testGlobalMetricValidation {
+    [self initializePinpointWithConfiguration:[self getDefaultAWSPinpointConfiguration] forceCreate:NO];
     @try {
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wnonnull"